Renamed arrays.py to arrays_kernels.py
